refactor: increased the width of add file dialog box + moved ListItem to a new file in the shared package
