+++ frontend/app/src/main/java/com/example/thinkr/ui/shared/ListItem.kt
@@ -1,0 +1,54 @@
+package com.example.thinkr.ui.shared
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.thinkr.R
+import com.example.thinkr.domain.model.DocumentItem
+import com.example.thinkr.ui.home.HomeScreenAction
+
+
+@Composable
+fun ListItem(item: DocumentItem, onAction: (HomeScreenAction) -> Unit = {}) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .clickable {
+                onAction(HomeScreenAction.DocumentItemClicked(documentItem = item))
+            }
+            .padding(vertical = 8.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        // Circular logo
+        Image(
+            painter = painterResource(id = R.drawable.document_placeholder_logo),
+            contentDescription = "Logo",
+            modifier = Modifier
+                .size(40.dp)
+                .clip(CircleShape)
+                .background(Color.Gray)
+        )
+
+        Spacer(modifier = Modifier.width(12.dp))
+
+        // Name
+        Text(text = item.name, fontSize = 18.sp, fontWeight = FontWeight.Medium)
+    }
+}
